+++ ci/kokoro/docker/build-in-docker-cmake.sh
@@ -179,6 +179,7 @@ if [[ "${TEST_INSTALL:-}" = "yes" ]]; then
         echo /var/tmp/staging/usr/local/include/google/cloud/bigtable/internal ; \
         echo /var/tmp/staging/usr/local/include/google/cloud/firestore ; \
         echo /var/tmp/staging/usr/local/include/google/cloud/grpc_utils ; \
+        echo /var/tmp/staging/usr/local/include/google/cloud/grpc_utils/internal ; \
         echo /var/tmp/staging/usr/local/include/google/cloud/internal ; \
         echo /var/tmp/staging/usr/local/include/google/cloud/spanner; \
         echo /var/tmp/staging/usr/local/include/google/cloud/storage ; \
