+++ google/cloud/bigtable/bigtable_client.bzl
@@ -19,7 +19,6 @@
 bigtable_client_hdrs = [
     "admin_client.h",
     "app_profile_config.h",
-    "async_operation.h",
     "async_row_reader.h",
     "cell.h",
     "client_options.h",
@@ -40,14 +39,12 @@ bigtable_client_hdrs = [
     "internal/async_bulk_apply.h",
     "internal/async_longrunning_op.h",
     "internal/async_poll_op.h",
-    "internal/async_read_stream_impl.h",
     "internal/async_retry_multi_page.h",
     "internal/async_retry_op.h",
     "internal/async_retry_unary_rpc.h",
     "internal/async_retry_unary_rpc_and_poll.h",
     "internal/bulk_mutator.h",
     "internal/client_options_defaults.h",
-    "internal/completion_queue_impl.h",
     "internal/common_client.h",
     "internal/conjunction.h",
     "internal/google_bytes_traits.h",
@@ -83,7 +80,6 @@ bigtable_client_srcs = [
     "app_profile_config.cc",
     "client_options.cc",
     "cluster_config.cc",
-    "completion_queue.cc",
     "data_client.cc",
     "expr.cc",
     "iam_binding.cc",
@@ -94,7 +90,6 @@ bigtable_client_srcs = [
     "instance_update_config.cc",
     "internal/async_bulk_apply.cc",
     "internal/bulk_mutator.cc",
-    "internal/completion_queue_impl.cc",
     "internal/common_client.cc",
     "internal/google_bytes_traits.cc",
     "internal/prefix_range_end.cc",
